requirements.txt
faiss-cpu==1.11.0
google_generativeai==0.8.5
python-dotenv==1.1.1
beautifulsoup4==4.13.4
flask==3.1.1